requirements.txt
torch>=2.0.0
pyyaml>=6.0
numpy>=1.21.0
pytest>=7.0
